+++ astrbot/core/platform/platform.py
@@ -10,7 +10,6 @@ from enum import Enum
 from typing import Any
 
 from astrbot.core.message.message_event_result import MessageChain
-from astrbot.core.utils.metrics import Metric
 
 from .astr_message_event import AstrMessageEvent
 from .astrbot_message import AstrBotMessage
@@ -163,9 +162,6 @@ class Platform(abc.ABC):
 
         异步方法。
         """
-        asyncio.create_task(
-            Metric.upload(msg_event_tick=1, adapter_name=self.meta().name)
-        )
 
     def commit_event(self, event: AstrMessageEvent) -> None:
         """提交一个事件到事件队列。"""