+++ astrbot/core/star/context.py
@@ -304,12 +304,12 @@ class Context:
             for k, v in kwargs.items()
             if k not in ["stream", "agent_hooks", "agent_context"]
         }
-        if request.func_tool and request.func_tool.get_tool("astrbot_file_read_tool"):
+        if request.func_tool and request.func_tool.get_tool("ldmbot_file_read"):
             other_kwargs.setdefault(
                 "tool_result_overflow_dir", get_astrbot_system_tmp_path()
             )
             other_kwargs.setdefault(
-                "read_tool", request.func_tool.get_tool("astrbot_file_read_tool")
+                "read_tool", request.func_tool.get_tool("ldmbot_file_read")
             )
 
         await agent_runner.reset(